Smoke_reminder/requirements.txt
python-telegram-bot[webhooks]==20.7
openai==1.18.0
python-dotenv
apscheduler
python-telegram-bot
python-dotenv
APScheduler==3.10.4
pytz==2023.3